move encoder in its own private package
