+++ daml-lf/archive/BUILD.bazel
@@ -6,7 +6,6 @@ load("//bazel_tools:pkg.bzl", "pkg_tar")
 load("//bazel_tools:proto.bzl", "proto_gen")
 load(
     "//bazel_tools:scala.bzl",
-    "da_scala_binary",
     "da_scala_library",
     "da_scala_test_suite",
     "lf_scalacopts",
