+++ daml-lf/encoder/src/main/scala/com/digitalasset/daml/lf/testing/archive/DamlLfEncoder.scala
@@ -1,30 +1,26 @@
 // Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.daml.lf.engine.assembler
+package com.digitalasset.daml.lf.testing.archive
 
 import java.nio.file.{Files, Paths}
 
-import com.digitalasset.daml.lf.archive.Encode
-import com.digitalasset.daml.lf.archive.EncodeV1.EncodeError
 import com.digitalasset.daml.lf.data.Ref
 import com.digitalasset.daml.lf.language.{Ast, LanguageMajorVersion, LanguageVersion}
+import com.digitalasset.daml.lf.testing.archive.EncodeV1.EncodeError
 import com.digitalasset.daml.lf.testing.parser.{ParserParameters, parseModules}
 import com.digitalasset.daml.lf.validation.Validation
 
 import scala.io.Source
 import scala.util.control.NonFatal
 
-object DamlLfAssembler extends App {
+private[digitalasset] object DamlLfEncoder extends App {
 
   private def error[X](message: String): X = {
     System.err.println(message)
     System.exit(1)
     throw new Error("You should not get this error")
   }
-
-  private def assert[X](e: Either[String, X]) =
-    e.fold(error, identity)
 
   private val pkgId = Ref.PackageId.assertFromString("-self-")
