+++ daml-lf/parser/src/main/scala/com/digitalasset/daml/lf/testing/parser/Implicits.scala
@@ -7,7 +7,7 @@ package parser
 import com.digitalasset.daml.lf.data.{Decimal, Ref}
 import com.digitalasset.daml.lf.language.Ast.{Expr, Kind, Package, Type}
 
-object Implicits {
+private[digitalasset] object Implicits {
 
   implicit val defaultParserParameters: ParserParameters[this.type] = ParserParameters(
     defaultPackageId,
