+++ daml-lf/parser/src/main/scala/com/digitalasset/daml/lf/testing/parser/ParserError.scala
@@ -3,12 +3,15 @@
 
 package com.digitalasset.daml.lf.testing.parser
 
-sealed abstract class ParserError(val description: String) extends RuntimeException(description)
+private[digitalasset] sealed abstract class ParserError(val description: String)
+    extends RuntimeException(description)
 
-final case class LexingError(override val description: String) extends ParserError(description) {
+private[digitalasset] final case class LexingError(override val description: String)
+    extends ParserError(description) {
   override def toString: String = s"ParsingError($description)"
 }
 
-final case class ParsingError(override val description: String) extends ParserError(description) {
+private[digitalasset] final case class ParsingError(override val description: String)
+    extends ParserError(description) {
   override def toString: String = s"ParsingError($description)"
 }
