+++ daml-lf/parser/src/main/scala/com/digitalasset/daml/lf/testing/parser/ParserParameters.scala
@@ -6,7 +6,7 @@ package com.digitalasset.daml.lf.testing.parser
 import com.digitalasset.daml.lf.data.Ref.PackageId
 import com.digitalasset.daml.lf.language.LanguageVersion
 
-case class ParserParameters[P](
+private[digitalasset] case class ParserParameters[P](
     defaultPackageId: PackageId,
     languageVersion: LanguageVersion
 )
